Fix `printf' format strings.
* src/block-indexer-chk.c, src/block-indexer-hash.c,
  src/chopper-anchor-based.c, src/cipher.c,
  src/filter-lzo-common.c, src/filter-lzo-unzip.c, src/filter-lzo-zip.c,
  src/filter-zip-push-pull.c, src/indexer-tree.c, src/logs.c,
  src/store-dummy.c, src/store-smart.c, src/store-stat.c,
  src/store-sunrpc.c, src/sunrpc-tls.c,
  tests/features/chopper-anchor-based.c, tests/features/filter-zip.c,
  tests/interfaces/choppers.c, tests/interfaces/cipher.c,
  tests/interfaces/class-lookup.c, tests/interfaces/filters.c,
  tests/interfaces/indexers.c, tests/interfaces/stores.c,
  utils/chop-archiver.c, utils/chop-show-anchors.c,
  utils/chop-show-similarities.c, utils/chop-store-convert.c,
  utils/chop-store-discover.c, utils/chop-store-list.c: Use `%zu' when
  using `printf ()' with `size_t' items.  Use `%l' in cases where the
  integer really is a `long'.

+++ src/store-sunrpc.c
@@ -45,6 +45,7 @@
 #endif
 
 #include <string.h>
+#include <stdint.h>
 #include <errno.h>
 
 
@@ -364,7 +365,7 @@ sunrpc_block_store_open (const char *host, unsigned port,
 	  return CHOP_INVALID_ARG;
 	}
 
-      gnutls_transport_set_ptr (session, (gnutls_transport_ptr_t)endpoint);
+      gnutls_transport_set_ptr (session, (gnutls_transport_ptr_t)(intptr_t) endpoint);
 
       /* Invoke the user-provided session initializer.  */
       err = init_session (session, closure);
